using BlockTracker (still under StateStorage at this point) to wait for a block

+++ config/config.go
@@ -25,7 +25,7 @@ type NodeConfig interface {
 
 	// state storage
 	StateHistoryRetentionInBlockHeights() uint64
-	QuerySyncGraceBlockDist() uint64
+	QuerySyncGraceBlockDist() uint16
 
 	// consensus context
 	BelowMinimalBlockDelayMillis() uint32
